% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{datareq_params}
\alias{datareq_params}
\title{Data Request Parameters}
\usage{
datareq_params(
  reqname = character(),
  research_fields = TRUE,
  prioritize_estimates = FALSE,
  filters = list()
)
}
\arguments{
\item{reqname}{Request name. Used as an identifier and for caching.}

\item{research_fields}{Should research fields be pulled? Defaults to TRUE}

\item{prioritize_estimates}{Should one priority estimate per study be returned (TRUE), or all estimates (FALSE)? Defaults to FALSE.}

\item{filters}{Named list, with filter names and a list of allowed options for that filter}
}
\description{
Creates an object specifying parameters for a request for SeroTracker data.
}
\examples{
fs_males_params <- datareq_params(reqname = "france_spain_males",
                                  research_fields = TRUE,
                                  prioritize_estimates = TRUE,
                                  filters = list(country = list("France", "Spain"),
                                                 sex = list("Male")))
}
\seealso{
\link{get_data} which takes these params as input
}
\keyword{import}
