% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{enforce_dtypes}
\alias{enforce_dtypes}
\title{Enforce Data Types}
\usage{
enforce_dtypes(tbl, dtypes, allcols = TRUE)
}
\arguments{
\item{tbl}{tibble to be dtyped}

\item{dtypes}{named list of character vectors: schema to be used for datatyping
names can include any of logical, double, int, date, factor, list
character vectors should be the names of all fields which should have that dtype}

\item{allcols}{logical: whether the whole tibble is being retyped,
in which case the function checks to ensure that there is a dtype provided
for every column}
}
\description{
Enforces data types on a tibble using a schema that maps feature names to dtypes.
}
\examples{
fs_males_tbl_cleaned <- enforce_dtypes(fs_males_tbl, serotracker_dtypes, TRUE)
}
\keyword{clean}
