% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{enforce_st_dtypes}
\alias{enforce_st_dtypes}
\title{Enforce SeroTracker Data Types}
\usage{
enforce_st_dtypes(tbl)
}
\arguments{
\item{tbl}{tibble to be dtyped}
}
\description{
Enforces SeroTracker data types on a tibble using a predefined schema.
}
\examples{
\dontrun{
fs_males_tbl_cleaned <- enforce_st_dtypes(fs_males_tbl)
}
}
\seealso{
wrapper method for \link{enforce_dtypes}
}
\keyword{clean}
