% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{filter_options}
\alias{filter_options}
\title{Filter Options}
\usage{
filter_options(server = "prod")
}
\arguments{
\item{server}{whether to use the production server ('prod') or a local dev server ('dev')}
}
\description{
Retrieve filter options from the data_provider/filter_options endpoint
}
\examples{
\dontrun{
filter_options()
}
}
\seealso{
Calls \link{retrieve_data}
}
\keyword{import}
