% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(params, import_new_data, path_to_cache_folder, server = "prod")
}
\arguments{
\item{params}{\link{datareq_params} object itemizing request parameters}

\item{import_new_data}{logical: requests data from server if TRUE, from cache if FALSE
Note that if import_new_data is TRUE, data will be repeatedly requested from server
even if a cache is present, and that cache will be overwritten
import_new_data must be set to FALSE to use cached data}

\item{path_to_cache_folder}{path to cache new data to, or read cache from}

\item{server}{whether to use the production server ('prod') or a local dev server ('dev')}
}
\description{
Gets serosurvey data from a cache or the SeroTracker servers.
If new data is requested, it is cached to the \code{path_to_cache_folder}.
Otherwise, new data is loaded from the \code{path_to_cache_folder}.
}
\examples{
\dontrun{
fs_males_tbl <- retrieve_data(fs_males_params,
                              TRUE,
                              '.cache')
}
}
\seealso{
calls \link{retrieve_data}
}
\keyword{import}
