% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve Data}
\usage{
retrieve_data(endpoint, params = NA)
}
\arguments{
\item{endpoint}{endpoint to be requested; "records" or "filter_options"}

\item{params}{\link{serosurvr:datareq_params} object itemizing request parameters}
}
\description{
Request SeroTracker data from a data_provider endpoint
}
\examples{
retrieve_data(fs_males_params)
}
\seealso{
\link{serosurvr:get_data} calls this function
}
\keyword{import}
