% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_dtypes.R
\docType{data}
\name{st_dtypes}
\alias{st_dtypes}
\title{Data types for all SeroTracker database elements.}
\format{
A named list of character vectors The name corresponds to the data
type (e.g., int), and the vector with that name enumerates all fields which
should have that data type.
}
\usage{
st_dtypes
}
\description{
Maps SeroTracker database elements to data types.
}
\keyword{datasets}
