% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{datareq_params}
\alias{datareq_params}
\title{Data Request Parameters}
\usage{
datareq_params(
  reqname = character(),
  research_fields = TRUE,
  prioritize_estimates = FALSE,
  prioritize_estimates_mode = "analysis_dynamic",
  columns = NULL,
  sampling_start_date = NULL,
  sampling_end_date = NULL,
  publication_start_date = NULL,
  publication_end_date = NULL,
  filters = list()
)
}
\arguments{
\item{reqname}{Request name. Used as an identifier and for caching.}

\item{research_fields}{Should research fields be pulled? Defaults to TRUE}

\item{prioritize_estimates}{Should one priority estimate per study be returned (TRUE), or all estimates (FALSE)? Defaults to FALSE.}

\item{prioritize_estimates_mode}{What mode should be used to prioritize estimates? Defaults to 'analysis_dynamic'; other options are 'analysis_static' and 'dashboard'}

\item{columns}{List or empty; if not empty, returns only the specified columns}

\item{sampling_start_date}{Filter results by sampling_start_date; format: YYYY-MM-DD}

\item{sampling_end_date}{Filter results by sampling_end_date; format: YYYY-MM-DD}

\item{publication_start_date}{Filter results by publication_start_date; format: YYYY-MM-DD}

\item{publication_end_date}{Filter results by publication_end_date; format: YYYY-MM-DD}

\item{filters}{Named list, with filter names and a list of allowed options for that filter}
}
\description{
Creates an object specifying parameters for a request for SeroTracker data.
}
\examples{
fs_males_params <- datareq_params(reqname = "france_spain_males",
                                  research_fields = TRUE,
                                  prioritize_estimates = TRUE,
                                  filters = list(country = list("France", "Spain"),
                                                 sex = list("Male")))
}
\seealso{
\link{get_data} which takes these params as input
}
\keyword{import}
